Added Player2 Script and updated all the interactables such that they can interact with both the player
