+++ SNAKE 2D/Assets/Scripts/Interactable/Food/Food.cs
@@ -11,10 +11,18 @@ public class Food : Interactable
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.gameObject.GetComponent<SnakeHeadMovement>() != null)
+        if (collision.gameObject.GetComponent<Player1>() != null)
         {
-            collision.gameObject.GetComponent<SnakeHeadMovement>().Grow();
-            collision.GetComponent<SnakeHeadMovement>().IncreaseScore(scorePoints);
+            collision.gameObject.GetComponent<Player1>().Grow();
+            collision.GetComponent<Player1>().IncreaseScore(scorePoints);
+            timer = respawnTime;
+            SpawnItem();
+        }
+
+        if (collision.gameObject.GetComponent<Player2>() != null)
+        {
+            collision.gameObject.GetComponent<Player2>().Grow();
+            collision.GetComponent<Player2>().IncreaseScore(scorePoints);
             timer = respawnTime;
             SpawnItem();
         }
