+++ SNAKE 2D/Assets/Scripts/Interactable/PowerUps/SpeedBoost.cs
@@ -15,18 +15,29 @@ public class SpeedBoost : PowerUps
     {
         if (powerTimer <= 0)
         {
+
             powerTimer = powerWearOffTime;
-            SnakeHeadMovement.Instance.ResetMoveRate();
+            Player1.Instance.ResetMoveRate();
+            Player2.Instance.ResetMoveRate();
         }
     }
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.GetComponent<SnakeHeadMovement>() != null)
+        if (collision.GetComponent<Player1>() != null)
         {
             powerTimer = powerWearOffTime;
-            collision.GetComponent<SnakeHeadMovement>().moveRate = boostSpeed;
+            Player2.Instance.ResetMoveRate();
+            collision.GetComponent<Player1>().moveRate = boostSpeed;
             HideItem();
         }
+        if (collision.GetComponent<Player2>() != null)
+        {
+            powerTimer = powerWearOffTime;
+            Player1.Instance.ResetMoveRate();
+            collision.GetComponent<Player2>().moveRate = boostSpeed;
+            HideItem();
+        }
+
     }
 }
