+++ SNAKE 2D/Assets/Scripts/Snake/Player1.cs
@@ -1,0 +1,80 @@
+using UnityEngine;
+
+public class Player1 : SnakeMovement
+{
+
+    static Player1 instance;
+    public static Player1 Instance { get { return instance; } }
+
+    private void Awake()
+    {
+        instance = this;
+    }
+
+    protected override void Update()
+    {
+        base.Update();
+        PlayerInput();
+    }
+
+    private void PlayerInput()
+    {
+        if (Input.GetKeyDown(KeyCode.W) && !BodyAt(transform.position.x, transform.position.y + 1))
+        {
+            direction = Vector2.up;
+        }
+        else if (Input.GetKeyDown(KeyCode.S) && !BodyAt(transform.position.x, transform.position.y - 1))
+        {
+            direction = Vector2.down;
+        }
+        else if (Input.GetKeyDown(KeyCode.D) && !BodyAt(transform.position.x + 1, transform.position.y))
+        {
+            direction = Vector2.right;
+        }
+        else if (Input.GetKeyDown(KeyCode.A) && !BodyAt(transform.position.x - 1, transform.position.y))
+        {
+            direction = Vector2.left;
+        }
+    }
+
+    public override void IncreaseScore(int value)
+    {
+        score += value * ScoreBooster.Instance.player1ScoreBoost;
+
+        GameManager.instance.UpdateScore(score);
+        GameManager.instance.UpdateLength(snakePositionsList.Count);
+    }
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+
+        if (collision.GetComponent<BodyData>() != null)
+        {
+            if (collision.GetComponent<BodyData>().snakeHead == this.gameObject && !immune)
+            {
+                GameManager.instance.GameOver();
+            }
+            if (collision.GetComponent<BodyData>().snakeHead != this.gameObject && !immune)
+            {
+                Debug.Log("Player2 Lost!!");
+            }
+
+        }
+
+
+    }
+
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.GetComponent<Player2>() != null &&
+            collision.gameObject.GetComponent<Player2>().GetSnakeLength() > this.GetSnakeLength())
+        {
+            Debug.Log("Player2 Win");
+        }
+        else if (collision.gameObject.GetComponent<Player2>() != null &&
+            collision.gameObject.GetComponent<Player2>().GetSnakeLength() < this.GetSnakeLength())
+        {
+            Debug.Log("Player1 Win");
+        }
+    }
+}
+
