+++ SNAKE 2D/Assets/Scripts/Snake/Player1.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 259e70dac27f3be4d94c131901a91812
+guid: cf8161c322d324f4aaee357d81fccaa2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
