+++ SNAKE 2D/Assets/Scripts/Snake/Player2.cs
@@ -1,0 +1,69 @@
+using UnityEngine;
+
+public class Player2 : SnakeMovement
+{
+
+    private static Player2 instance;
+    public static Player2 Instance { get { return instance; } }
+
+    private void Awake()
+    {
+        instance = this;
+    }
+
+    protected override void Update()
+    {
+        base.Update();
+        PlayerInput();
+
+    }
+
+
+    private void PlayerInput()
+    {
+        if (Input.GetKeyDown(KeyCode.UpArrow) && !BodyAt(transform.position.x, transform.position.y + 1))
+        {
+            Debug.Log("Up arrow");
+            direction = Vector2.up;
+        }
+        else if (Input.GetKeyDown(KeyCode.DownArrow) && !BodyAt(transform.position.x, transform.position.y - 1))
+        {
+            direction = Vector2.down;
+        }
+        else if (Input.GetKeyDown(KeyCode.RightArrow) && !BodyAt(transform.position.x + 1, transform.position.y))
+        {
+            direction = Vector2.right;
+        }
+        else if (Input.GetKeyDown(KeyCode.LeftArrow) && !BodyAt(transform.position.x - 1, transform.position.y))
+        {
+            direction = Vector2.left;
+        }
+    }
+
+    public override void IncreaseScore(int value)
+    {
+        score += value * ScoreBooster.Instance.player2ScoreBoost;
+
+        GameManager.instance.UpdateScore(score);
+        GameManager.instance.UpdateLength(snakePositionsList.Count);
+    }
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+
+        if (collision.GetComponent<BodyData>() != null)
+        {
+            if (collision.GetComponent<BodyData>().snakeHead == this.gameObject && !immune)
+            {
+                GameManager.instance.GameOver();
+            }
+            if (collision.GetComponent<BodyData>().snakeHead != this.gameObject && !immune)
+            {
+                Debug.Log("Player1 Lost!!");
+            }
+
+        }
+    }
+
+
+}
