+++ SNAKE 2D/Assets/Scripts/Snake/Player2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc0bf1be75bebe44895b247124203305
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
